ai_agent_complex/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]>=0.31.1
pydantic>=2.7.0
pydantic-settings>=2.5.2
httpx>=0.27.1
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10,<0.1.0
langchain-core>=0.1.40,<0.2.0
langchain-openai>=0.0.2,<0.1.0
langgraph>=0.0.12,<0.1.0
openai>=1.12.0
python-dotenv==1.0.0


# Observability
prometheus_client>=0.19.0
# MCP (Model Context Protocol) client - Official SDK from GitHub
git+https://github.com/modelcontextprotocol/python-sdk

# RAG dependencies
chromadb==0.4.22
tiktoken==0.5.2
python-multipart>=0.0.9
langchain-chroma>=0.1.0,<0.2.0

# Teaching Memory Lab dependencies
aiosqlite==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
